app/requirements.txt
fastapi
uvicorn
networkx
matplotlib
numpy
pydantic
pulp
google-generativeai
python-dotenv
Pillow
